Split of data reconstruction to a separate file, showing whole reconstruction process if DEBUG is true
